requirements.txt
functions-framework
flask
psycopg2-binary
google-cloud-aiplatform
google-auth